feat: API для работы с событиями, запросами на участие

+++ stats-service/service/src/main/java/ru/practicum/stats/model/ViewStats.java
@@ -12,5 +12,5 @@ import lombok.ToString;
 public class ViewStats {
     private String app;
     private String uri;
-    private long hits;
+    private int hits;
 }
